// temper/game-crafting-addon/src/potion-maker/types.ts
export interface BagSlot {
  bagId: number
  slotIndex: number
  stack: number
}

export interface Reagent {
  traits: Record<string, boolean>
  itemId: number
  itemLink?: string
  matching?: number[]
}

export interface MatchResult {
  matched: Record<string, TraitEffect>
  passSingleTrait: boolean
  count: number
}

export interface Ingredient {
  name: string
  itemId: number
  level: number
  icon: string
  stack: number
  pack: BagSlot[]
  traits: Record<string, TraitEffect>
  itemLink?: string
  matchIngredient1: (this: void, self: Ingredient, passiveIngredient: Ingredient) => MatchResult
  matchIngredient: (this: void, self: Ingredient, passiveIngredient: Ingredient) => MatchResult
  ResetStack: (this: void, self: Ingredient) => void
}

export interface IngredientFactory {
  name: string
  itemId: number
  level: number
  icon: string
  new: (this: void, o: Partial<Ingredient> & { itemId: number }) => Ingredient
  solvent: (this: void, o: Partial<Ingredient>) => Ingredient
}

export interface Potion {
  name: string
  searchName: string
  upperName: string
  solvent: Ingredient | undefined
  ingredients: Ingredient[]
  traits: Record<string, TraitEffect>
  numTraits: number
  qualityColor: string | ZoColorDef
  quantity: number
  itemLink: string
  itemId: string | number
  samePotionId: string | number
  sameTraitsId: string | number
}